+++ src/main/java/fll/report/awards/EditCategoriesAwarded.java
@@ -19,6 +19,7 @@ import java.util.stream.Collectors;
 
 import javax.sql.DataSource;
 
+import org.apache.commons.lang3.StringUtils;
 import org.checkerframework.checker.nullness.qual.Nullable;
 
 import fll.TournamentLevel;